Fixes #1157 by renaming filelock to filelock_brat
Fixed the filename which was causing the problems.
Fixed imports of the file.
